+++ Sources/MapboxMaps/Foundation/MapboxMap.swift
@@ -536,9 +536,7 @@ public final class MapboxMap: MapboxMapProtocol {
         let expected = __map.setBoundsFor(MapboxCoreMaps.CameraBoundsOptions(options))
 
         if expected.isError() {
-            // swiftlint:disable force_cast
-            throw MapError(coreError: expected.error as! NSString)
-            // swiftlint:enable force_cast
+            throw MapError(coreError: expected.error)
         }
     }
 
@@ -576,7 +574,7 @@ public final class MapboxMap: MapboxMapProtocol {
     }
 
     internal func pointIsAboveHorizon(_ point: CGPoint) -> Bool {
-        guard case .mercator = try? mapProjection() else {
+        guard style.projection.name == .mercator else {
             return false
         }
         let topMargin = 0.04 * size.height
@@ -951,7 +949,7 @@ extension MapboxMap {
                               sourceLayerId: sourceLayerId,
                               featureId: featureId,
                               callback: coreAPIClosureAdapter(for: callback,
-                                                              type: NSDictionary.self,
+                                                              type: AnyObject.self,
                                                               concreteErrorType: MapError.self))
     }
 
@@ -982,66 +980,36 @@ extension MapboxMap {
 
     internal func addViewAnnotation(withId id: String, options: ViewAnnotationOptions) throws {
         let expected = __map.addViewAnnotation(forIdentifier: id, options: MapboxCoreMaps.ViewAnnotationOptions(options))
-        if expected.isError(), let reason = expected.error as? NSString {
+        if expected.isError(), let reason = expected.error {
             throw MapError(coreError: reason)
         }
     }
 
     internal func updateViewAnnotation(withId id: String, options: ViewAnnotationOptions) throws {
         let expected = __map.updateViewAnnotation(forIdentifier: id, options: MapboxCoreMaps.ViewAnnotationOptions(options))
-        if expected.isError(), let reason = expected.error as? NSString {
+        if expected.isError(), let reason = expected.error {
             throw MapError(coreError: reason)
         }
     }
 
     internal func removeViewAnnotation(withId id: String) throws {
         let expected = __map.removeViewAnnotation(forIdentifier: id)
-        if expected.isError(), let reason = expected.error as? NSString {
+        if expected.isError(), let reason = expected.error {
             throw MapError(coreError: reason)
         }
     }
 
     internal func options(forViewAnnotationWithId id: String) throws -> ViewAnnotationOptions {
         let expected = __map.getViewAnnotationOptions(forIdentifier: id)
-        if expected.isError(), let reason = expected.error as? NSString {
+        if expected.isError(), let reason = expected.error {
             throw MapError(coreError: reason)
         }
-        guard let options = expected.value as? MapboxCoreMaps.ViewAnnotationOptions else {
+        guard let options = expected.value else {
             fatalError("Failed to unwrap ViewAnnotationOptions")
         }
         return ViewAnnotationOptions(options)
     }
 
-}
-
-// MARK: - MapProjection
-
-extension MapboxMap {
-    /// Errors related to MapProjection API
-    @_spi(Experimental) public enum MapProjectionError: Error {
-        case unsupportedProjection
-    }
-
-    /// Set map projection for Mapbox map.
-    /// - Parameter mode: The `MapProjection` to be used by the map.
-    /// - Throws: Errors during encoding or `MapProjectionError.unsupportedProjection` if the supplied projection is not compatible with the SDK.
-    @_spi(Experimental) public func setMapProjection(_ mapProjection: MapProjection) throws {
-        try __map.setMapProjectionForProjection(mapProjection.toJSON())
-    }
-
-    /// Get current map projection for Mapbox map.
-    ///
-    /// Please note that even if MapboxMap is configured to use `MapProjection.globe`
-    /// starting from `GlobeMapProjection.transitionZoomLevel` and above
-    /// this method will return `MapProjection.mercator`.
-    ///
-    /// - Returns:
-    ///     `MapProjection` map is using.
-    /// - Throws: Errors during decoding
-    @_spi(Experimental) public func mapProjection() throws -> MapProjection {
-        let data = try JSONSerialization.data(withJSONObject: __map.getMapProjection(), options: [])
-        return try JSONDecoder().decode(MapProjection.self, from: data)
-    }
 }
 
 // MARK: - Testing only!