+++ Sources/MapboxMaps/Offline/OfflineManager+MapboxMaps.swift
@@ -119,6 +119,6 @@ extension OfflineManager {
 
 private func offlineManagerClosureAdapter<T, ObjCType>(
     for closure: @escaping (Result<T, Error>) -> Void,
-    type: ObjCType.Type) -> ((Expected<AnyObject, AnyObject>?) -> Void) where ObjCType: AnyObject {
+    type: ObjCType.Type) -> ((Expected<ObjCType, StylePackError.CoreErrorType>?) -> Void) where ObjCType: AnyObject {
     return coreAPIClosureAdapter(for: closure, type: type, concreteErrorType: StylePackError.self)
 }