+++ Tests/MapboxMapsTests/Foundation/MapboxMapTests.swift
@@ -1,5 +1,5 @@
 import XCTest
-@_spi(Experimental) @testable import MapboxMaps
+@testable import MapboxMaps
 @_implementationOnly import MapboxCoreMaps_Private
 
 final class MapboxMapTests: XCTestCase {
@@ -212,22 +212,6 @@ final class MapboxMapTests: XCTestCase {
         mapboxMap.endGesture()
 
         XCTAssertFalse(mapboxMap.__testingMap.isGestureInProgress())
-    }
-
-    func testSetMapProjection() {
-        XCTAssertEqual(mapboxMap.__testingMap.getMapProjection() as? [String: String], ["name": "mercator"])
-        try? mapboxMap.setMapProjection(.globe())
-        XCTAssertEqual(mapboxMap.__testingMap.getMapProjection() as? [String: String], ["name": "globe"])
-    }
-
-    func testGetMapProjection() {
-        mapboxMap.__testingMap.setMapProjectionForProjection(["name": "mercator"])
-        var projection = try? mapboxMap.mapProjection()
-        XCTAssertEqual(projection, .mercator())
-
-        mapboxMap.__testingMap.setMapProjectionForProjection(["name": "globe"])
-        projection = try? mapboxMap.mapProjection()
-        XCTAssertEqual(projection, .globe())
     }
 
     func testSubscribe() throws {