+++ CHANGELOG.md
@@ -22,6 +22,9 @@ Mapbox welcomes participation and contributions from everyone.
 * Update `DefaultViewportTransition` to solve the moving target problem. ([#1245](https://github.com/mapbox/mapbox-maps-ios/pull/1245))
 * Increase deceleration cutoff threshold from 20 to 35 to prevent camera changes
  after animation stops. ([#1244](https://github.com/mapbox/mapbox-maps-ios/pull/1244))
+* Update to MapboxCoreMaps 10.5.0-beta.1 and MapboxCommon 21.3.0-beta.2. ([]())
+* Update to MapboxCoreMaps 10.5.0-beta.1 and MapboxCommon 21.3.0-beta.2. ([#1235](https://github.com/mapbox/mapbox-maps-ios/pull/1235))
+* API for using globe projection has been moved to `Style.setProjection(_:)` and `Style.projection` and is no longer experimental. ([#1235](https://github.com/mapbox/mapbox-maps-ios/pull/1235)) 
 
 ## 10.4.1 - March 28, 2022